refactor : 테스트클래스 추가
