+++ src/main/java/calc/Calculation.java
@@ -20,8 +20,8 @@ public class Calculation {
         return result;
     }
 
-    public int calculate(int a, int b, String operation) {
-        return Operation.findByOperation(operation).calc(a, b);
+    public int calculate(int operand1, int operand2, String operation) {
+        return Operation.findByOperation(operation).calc(operand1, operand2);
     }
 
 }
